+++ packages/expo-updates/ios/Tests/FileDownloaderManifestParsingSpec.swift
@@ -42,16 +42,176 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
         
         let bodyData = classicJSON.data(using: .utf8)!
         
-        var resultUpdateManifest: Update? = nil
+        var resultUpdateResponse: UpdateResponse? = nil
         var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
         } errorBlock: { error in
           errorOccurred = error
         }
         
         expect(errorOccurred).to(beNil())
-        expect(resultUpdateManifest).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart?.updateManifest.manifest.isVerified()) == false
+      }
+      
+      it("multipart body") {
+        let config = UpdatesConfig.config(fromDictionary: [
+          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
+        ])
+        let downloader = FileDownloader(config: config)
+        
+        let boundary = "blah"
+        let contentType = "multipart/mixed; boundary=\(boundary)"
+        let response = HTTPURLResponse(
+          url: URL(string: "https://exp.host/@test/test")!,
+          statusCode: 200,
+          httpVersion: "HTTP/1.1",
+          headerFields: ["content-type": contentType]
+        )!
+        
+        let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
+          boundary: boundary,
+          manifest: classicJSON,
+          manifestSignature: nil,
+          certificateChain: nil,
+          directive: TestHelper.testDirectiveNoUpdateAvailable,
+          directiveSignature: nil
+        )
+        
+        var resultUpdateResponse: UpdateResponse? = nil
+        var errorOccurred: (any Error)? = nil
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
+        } errorBlock: { error in
+          errorOccurred = error
+        }
+        
+        expect(errorOccurred).to(beNil())
+        expect(resultUpdateResponse).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart?.updateManifest.manifest.isVerified()) == false
+        
+        expect(resultUpdateResponse?.directiveUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.directiveUpdateResponsePart?.updateDirective is NoUpdateAvailableUpdateDirective) == true
+      }
+      
+      it("multipart body only directive") {
+        let config = UpdatesConfig.config(fromDictionary: [
+          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
+        ])
+        let downloader = FileDownloader(config: config)
+        
+        let boundary = "blah"
+        let contentType = "multipart/mixed; boundary=\(boundary)"
+        let response = HTTPURLResponse(
+          url: URL(string: "https://exp.host/@test/test")!,
+          statusCode: 200,
+          httpVersion: "HTTP/1.1",
+          headerFields: ["content-type": contentType]
+        )!
+        
+        let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
+          boundary: boundary,
+          manifest: nil,
+          manifestSignature: nil,
+          certificateChain: nil,
+          directive: TestHelper.testDirectiveNoUpdateAvailable,
+          directiveSignature: nil
+        )
+        
+        var resultUpdateResponse: UpdateResponse? = nil
+        var errorOccurred: (any Error)? = nil
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
+        } errorBlock: { error in
+          errorOccurred = error
+        }
+        
+        expect(errorOccurred).to(beNil())
+        expect(resultUpdateResponse).notTo(beNil())
+        
+        expect(resultUpdateResponse?.manifestUpdateResponsePart).to(beNil())
+        
+        expect(resultUpdateResponse?.directiveUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.directiveUpdateResponsePart?.updateDirective is NoUpdateAvailableUpdateDirective) == true
+      }
+      
+      it("multipart body only directive v0 compatibility mode") {
+        let config = UpdatesConfig.config(fromDictionary: [
+          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
+          UpdatesConfig.EXUpdatesConfigEnableExpoUpdatesProtocolV0CompatibilityModeKey: true
+        ])
+        let downloader = FileDownloader(config: config)
+        
+        let boundary = "blah"
+        let contentType = "multipart/mixed; boundary=\(boundary)"
+        let response = HTTPURLResponse(
+          url: URL(string: "https://exp.host/@test/test")!,
+          statusCode: 200,
+          httpVersion: "HTTP/1.1",
+          headerFields: ["content-type": contentType]
+        )!
+        
+        let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
+          boundary: boundary,
+          manifest: nil,
+          manifestSignature: nil,
+          certificateChain: nil,
+          directive: TestHelper.testDirectiveNoUpdateAvailable,
+          directiveSignature: nil
+        )
+        
+        var resultUpdateResponse: UpdateResponse? = nil
+        var errorOccurred: (any Error)? = nil
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
+        } errorBlock: { error in
+          errorOccurred = error
+        }
+        
+        expect(errorOccurred?.localizedDescription) == "Multipart response missing manifest part. Manifest is required in version 0 of the expo-updates protocol. This may be due to the update being a rollback or other directive."
+        expect(resultUpdateResponse).to(beNil())
+      }
+      
+      it("multipart body no relevant parts") {
+        let config = UpdatesConfig.config(fromDictionary: [
+          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
+        ])
+        let downloader = FileDownloader(config: config)
+        
+        let boundary = "blah"
+        let contentType = "multipart/mixed; boundary=\(boundary)"
+        let response = HTTPURLResponse(
+          url: URL(string: "https://exp.host/@test/test")!,
+          statusCode: 200,
+          httpVersion: "HTTP/1.1",
+          headerFields: ["content-type": contentType]
+        )!
+        
+        let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
+          boundary: boundary,
+          manifest: nil,
+          manifestSignature: nil,
+          certificateChain: nil,
+          directive: nil,
+          directiveSignature: nil,
+          extraneous: ""
+        )
+        
+        var resultUpdateResponse: UpdateResponse? = nil
+        var errorOccurred: (any Error)? = nil
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
+        } errorBlock: { error in
+          errorOccurred = error
+        }
+        
+        expect(errorOccurred).to(beNil())
+        expect(resultUpdateResponse).notTo(beNil())
+        
+        expect(resultUpdateResponse?.manifestUpdateResponsePart).to(beNil())
+        expect(resultUpdateResponse?.directiveUpdateResponsePart).to(beNil())
       }
       
       it("json body signed") {
@@ -76,50 +236,17 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
         
         let bodyData = modernJSON.data(using: .utf8)!
         
-        var resultUpdateManifest: Update? = nil
+        var resultUpdateResponse: UpdateResponse? = nil
         var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
         } errorBlock: { error in
           errorOccurred = error
         }
         
         expect(errorOccurred).to(beNil())
-        expect(resultUpdateManifest).notTo(beNil())
-      }
-      
-      it("multipart body") {
-        let config = UpdatesConfig.config(fromDictionary: [
-          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
-        ])
-        let downloader = FileDownloader(config: config)
-        
-        let boundary = "blah"
-        let contentType = "multipart/mixed; boundary=\(boundary)"
-        let response = HTTPURLResponse(
-          url: URL(string: "https://exp.host/@test/test")!,
-          statusCode: 200,
-          httpVersion: "HTTP/1.1",
-          headerFields: ["content-type": contentType]
-        )!
-        
-        let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
-          fromManifest: classicJSON,
-          withBoundary: boundary,
-          manifestSignature: nil,
-          certificateChain: nil
-        )
-        
-        var resultUpdateManifest: Update? = nil
-        var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
-        } errorBlock: { error in
-          errorOccurred = error
-        }
-        
-        expect(errorOccurred).to(beNil())
-        expect(resultUpdateManifest).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart?.updateManifest.manifest.isVerified()) == true
       }
       
       it("multipart body signed") {
@@ -144,22 +271,30 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
         )!
         
         let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
-          fromManifest: modernJSON,
-          withBoundary: boundary,
+          boundary: boundary,
+          manifest: modernJSON,
           manifestSignature: modernJSONSignature,
-          certificateChain: nil
+          certificateChain: nil,
+          directive: TestHelper.testDirectiveNoUpdateAvailable,
+          directiveSignature: TestHelper.testDirectiveNoUpdateAvailableSignature
         )
         
-        var resultUpdateManifest: Update? = nil
+        var resultUpdateResponse: UpdateResponse? = nil
         var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
         } errorBlock: { error in
           errorOccurred = error
         }
         
         expect(errorOccurred).to(beNil())
-        expect(resultUpdateManifest).notTo(beNil())
+        expect(resultUpdateResponse).notTo(beNil())
+        expect(resultUpdateResponse).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart?.updateManifest.manifest.isVerified()) == true
+        
+        expect(resultUpdateResponse?.directiveUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.directiveUpdateResponsePart?.updateDirective is NoUpdateAvailableUpdateDirective) == true
       }
       
       it("json body expects signed receives unsigned") {
@@ -183,50 +318,16 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
         
         let bodyData = modernJSON.data(using: .utf8)!
         
-        var resultUpdateManifest: Update? = nil
+        var resultUpdateResponse: UpdateResponse? = nil
         var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
         } errorBlock: { error in
           errorOccurred = error
         }
         
         expect(errorOccurred?.localizedDescription) == "No expo-signature header specified"
-        expect(resultUpdateManifest).to(beNil())
-      }
-      
-      it("json body signed unsigned request manifest signature optional") {
-        let config = UpdatesConfig.config(fromDictionary: [
-          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
-          UpdatesConfig.EXUpdatesConfigCodeSigningCertificateKey: modernJSONCertificate,
-          UpdatesConfig.EXUpdatesConfigCodeSigningMetadataKey: [:],
-          UpdatesConfig.EXUpdatesConfigCodeSigningAllowUnsignedManifestsKey: true
-        ])
-        let downloader = FileDownloader(config: config)
-        let contentType = "application/json"
-        let response = HTTPURLResponse(
-          url: URL(string: "https://exp.host/@test/test")!,
-          statusCode: 200,
-          httpVersion: "HTTP/1.1",
-          headerFields: [
-            "expo-protocol-version": "0",
-            "expo-sfv-version": "0",
-            "content-type": contentType,
-          ]
-        )!
-        
-        let bodyData = modernJSON.data(using: .utf8)!
-        
-        var resultUpdateManifest: Update? = nil
-        var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
-        } errorBlock: { error in
-          errorOccurred = error
-        }
-        
-        expect(errorOccurred).to(beNil())
-        expect(resultUpdateManifest).notTo(beNil())
+        expect(resultUpdateResponse).to(beNil())
       }
       
       it("multipart body signed certificate particular experience") {
@@ -254,23 +355,28 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
         )!
         
         let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
-          fromManifest: modernJSON,
-          withBoundary: boundary,
+          boundary: boundary,
+          manifest: modernJSON,
           manifestSignature: chainLeafSignature,
-          certificateChain: "\(leafCertificate)\(intermediateCertificate)"
+          certificateChain: "\(leafCertificate)\(intermediateCertificate)",
+          directive: TestHelper.testDirectiveNoUpdateAvailable,
+          directiveSignature: TestHelper.testDirectiveNoUpdateAvailableValidChainLeafSignature
         )
         
-        var resultUpdateManifest: Update? = nil
+        var resultUpdateResponse: UpdateResponse? = nil
         var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
         } errorBlock: { error in
           errorOccurred = error
         }
         
         expect(errorOccurred).to(beNil())
-        expect(resultUpdateManifest).notTo(beNil())
-        expect(resultUpdateManifest?.manifest.rawManifestJSON()["isVerified"] as? Bool) == true
+        expect(resultUpdateResponse).notTo(beNil())
+        expect(resultUpdateResponse?.manifestUpdateResponsePart?.updateManifest.manifest.isVerified()) == true
+        
+        expect(resultUpdateResponse?.directiveUpdateResponsePart).notTo(beNil())
+        expect(resultUpdateResponse?.directiveUpdateResponsePart?.updateDirective is NoUpdateAvailableUpdateDirective) == true
       }
       
       it("multipart body signed certificate particular experience incorrect experience in manifest") {
@@ -298,47 +404,140 @@ class FileDownloaderManifestParsingSpec : ExpoSpec {
         )!
         
         let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
-          fromManifest: manifestBodyIncorrectProjectId,
-          withBoundary: boundary,
+          boundary: boundary,
+          manifest: manifestBodyIncorrectProjectId,
           manifestSignature: validChainLeafSignatureIncorrectProjectId,
-          certificateChain: "\(leafCertificate)\(intermediateCertificate)"
+          certificateChain: "\(leafCertificate)\(intermediateCertificate)",
+          directive: nil,
+          directiveSignature: nil
         )
         
-        var resultUpdateManifest: Update? = nil
+        var resultUpdateResponse: UpdateResponse? = nil
         var errorOccurred: (any Error)? = nil
-        downloader.parseManifestResponse(response, withData: bodyData, database: database) { update in
-          resultUpdateManifest = update
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
         } errorBlock: { error in
           errorOccurred = error
         }
         
         expect(errorOccurred?.localizedDescription) == "Invalid certificate for manifest project ID or scope key"
-        expect(resultUpdateManifest).to(beNil())
+        expect(resultUpdateResponse).to(beNil())
+      }
+      
+      it("multipart body signed certificate particular experience incorrect experience in directive") {
+        let config = UpdatesConfig.config(fromDictionary: [
+          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
+          UpdatesConfig.EXUpdatesConfigCodeSigningCertificateKey: rootCertificate,
+          UpdatesConfig.EXUpdatesConfigCodeSigningMetadataKey: [
+            "keyid": "ca-root",
+          ],
+          UpdatesConfig.EXUpdatesConfigCodeSigningIncludeManifestResponseCertificateChainKey: true
+        ])
+        let downloader = FileDownloader(config: config)
+        
+        let boundary = "blah"
+        let contentType = "multipart/mixed; boundary=\(boundary)"
+        let response = HTTPURLResponse(
+          url: URL(string: "https://exp.host/@test/test")!,
+          statusCode: 200,
+          httpVersion: "HTTP/1.1",
+          headerFields: [
+            "expo-protocol-version": "0",
+            "expo-sfv-version": "0",
+            "content-type": contentType
+          ]
+        )!
+        
+        let bodyData = FileDownloaderManifestParsingSpec.mutlipartData(
+          boundary: boundary,
+          manifest: nil,
+          manifestSignature: nil,
+          certificateChain: "\(leafCertificate)\(intermediateCertificate)",
+          directive: TestHelper.testDirectiveNoUpdateAvailableIncorrectProjectId,
+          directiveSignature: TestHelper.testDirectiveNoUpdateAvailableValidChainLeafSignatureIncorrectProjectId
+        )
+        
+        var resultUpdateResponse: UpdateResponse? = nil
+        var errorOccurred: (any Error)? = nil
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
+        } errorBlock: { error in
+          errorOccurred = error
+        }
+        
+        expect(errorOccurred?.localizedDescription) == "Invalid certificate for directive project ID or scope key"
+        expect(resultUpdateResponse).to(beNil())
+      }
+      
+      it("json body signed unsigned request manifest signature optional") {
+        let config = UpdatesConfig.config(fromDictionary: [
+          UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test",
+          UpdatesConfig.EXUpdatesConfigCodeSigningCertificateKey: modernJSONCertificate,
+          UpdatesConfig.EXUpdatesConfigCodeSigningMetadataKey: [:],
+          UpdatesConfig.EXUpdatesConfigCodeSigningAllowUnsignedManifestsKey: true
+        ])
+        let downloader = FileDownloader(config: config)
+        let contentType = "application/json"
+        let response = HTTPURLResponse(
+          url: URL(string: "https://exp.host/@test/test")!,
+          statusCode: 200,
+          httpVersion: "HTTP/1.1",
+          headerFields: [
+            "expo-protocol-version": "0",
+            "expo-sfv-version": "0",
+            "content-type": contentType,
+          ]
+        )!
+        
+        let bodyData = modernJSON.data(using: .utf8)!
+        
+        var resultUpdateResponse: UpdateResponse? = nil
+        var errorOccurred: (any Error)? = nil
+        downloader.parseManifestResponse(response, withData: bodyData, database: database) { updateResponse in
+          resultUpdateResponse = updateResponse
+        } errorBlock: { error in
+          errorOccurred = error
+        }
+        
+        expect(errorOccurred).to(beNil())
+        expect(resultUpdateResponse).notTo(beNil())
       }
     }
   }
   
   private static func mutlipartData(
-    fromManifest manifest: String,
-    withBoundary boundary: String,
-    manifestSignature signature: String?,
-    certificateChain: String?
+    boundary: String,
+    manifest: String?,
+    manifestSignature: String?,
+    certificateChain: String?,
+    directive: String?,
+    directiveSignature: String?,
+    extraneous: String? = nil
   ) -> Data {
     var body = Data()
-    body.appendStringData("--\(boundary)\r\n")
-    body.appendStringData("Content-Type: application/json\r\n")
-    if let signature = signature {
-      body.appendStringData("expo-signature: \(signature)\r\n")
-    }
-    body.appendStringData("Content-Disposition: inline; name=\"manifest\"\r\n\r\n")
-    body.appendStringData(manifest)
-    body.appendStringData("\r\n")
-    if let certificateChain = certificateChain {
+    
+    func appendPart(name: String, contentType: String, bodyString: String, signature: String? = nil) {
       body.appendStringData("--\(boundary)\r\n")
-      body.appendStringData("Content-Type: application/x-pem-file\r\n")
-      body.appendStringData("Content-Disposition: inline; name=\"certificate_chain\"\r\n\r\n")
-      body.appendStringData(certificateChain)
+      body.appendStringData("Content-Type: application/\(contentType)\r\n")
+      if let signature = signature {
+        body.appendStringData("expo-signature: \(signature)\r\n")
+      }
+      body.appendStringData("Content-Disposition: inline; name=\"\(name)\"\r\n\r\n")
+      body.appendStringData(bodyString)
       body.appendStringData("\r\n")
+    }
+    
+    if let manifest = manifest {
+      appendPart(name: "manifest", contentType: "json", bodyString: manifest, signature: manifestSignature)
+    }
+    if let certificateChain = certificateChain {
+      appendPart(name: "certificate_chain", contentType: "x-pem-file", bodyString: certificateChain)
+    }
+    if let directive = directive {
+      appendPart(name: "directive", contentType: "json", bodyString: directive, signature: directiveSignature)
+    }
+    if let extraneous = extraneous {
+      appendPart(name: "extraneous", contentType: "text", bodyString: extraneous)
     }
     body.appendStringData("--\(boundary)--\r\n")
     return body